refactor: Gamer 인터페이스를 부모 클래스로 변경
