+++ src/main/java/domain/gamer/Dealer.java
@@ -1,0 +1,17 @@
+package domain.gamer;
+
+import domain.card.Hands;
+
+/**
+ *   class description
+ *
+ *   @author ParkDooWon, AnHyungJu  
+ */
+public class Dealer extends Gamer {
+	private String name;
+	private Hands hands;
+
+	public Dealer() {
+		super("딜러");
+	}
+}
