+++ src/main/java/domain/gamer/Gamer.java
@@ -1,18 +1,35 @@
 package domain.gamer;
 
 import domain.card.Card;
+import domain.card.Hands;
 
 /**
- *    게임 진행자 인터페이스입니다.
+ *    게임 참가자 부모 클래스입니다.
  *
  *    @author AnHyungJu, ParkDooWon
  */
-public interface Gamer {
-	boolean isBust();
+public class Gamer {
+	private String name;
+	private Hands hands;
 
-	void add(Card card);
+	public Gamer(String name) {
+		this.name = name;
+		this.hands = new Hands();
+	}
 
-	int scoreHands();
+	public void add(Card card) {
+		hands.add(card);
+	}
 
-	boolean isBlackjack();
+	public boolean isBust() {
+		return hands.calculateTotalScore() > Hands.BLACKJACK;
+	}
+
+	public int scoreHands() {
+		return hands.calculateTotalScore();
+	}
+
+	public boolean isBlackjack() {
+		return (hands.calculateTotalScore() == Hands.BLACKJACK) && hands.hasTwoCards();
+	}
 }
